refactor(font): change font data change bind and pick options
