+++ src/views/pageTypewriter/components/ConfigBar/ButtonFont.vue
@@ -1,58 +1,14 @@
 <script setup lang="ts">
 import { ref, computed } from "vue";
 import type { ComputedRef } from "vue";
-import type { FontGroupNode, FontNode } from "@/shared";
-import { useConfig } from "../../hooks";
+import type { FontNode } from "@/shared";
+import { useConfig, useFont } from "../../hooks";
 import ButtonBase from "./ButtonBase.vue";
 import ButtonDropdown from "./ButtonDropdown.vue";
 import ImgFontFamily from "./assets/font-family.svg";
-import ImgLogoGenshin from "./assets/logo-genshin.svg";
-import ImgLogoStarrail from "./assets/logo-starrail.svg";
-import ImgLogoZzz from "./assets/logo-zzz.svg";
-
-const fontOptions: FontGroupNode[] = [
-  {
-    id: "genshin",
-    label: "原神",
-    icon: ImgLogoGenshin,
-    displayWidgets: ["icon"],
-    children: [
-      { tag: "font-teyvat-regular", label: "提瓦特文字", abbr: "提瓦特文" },
-      { tag: "font-khaenriah-regular", label: "坎瑞亚文字", abbr: "坎瑞亚文" },
-      {
-        tag: "font-khaenriah-chasm",
-        label: "坎瑞亚文字 - 层岩巨渊变体",
-        abbr: "坎瑞亚文（层岩变体）",
-      },
-      { tag: "font-inazuma-regular", label: "稻妻文字", abbr: "稻妻文" },
-      { tag: "font-sumeru-regular", label: "须弥文字", abbr: "须弥文" },
-      { tag: "font-deshret-regular", label: "赤冠文字", abbr: "赤冠文" },
-    ],
-  },
-  {
-    id: "starrail",
-    label: "崩坏·星穹铁道",
-    icon: ImgLogoStarrail,
-    displayWidgets: ["icon", "label"],
-    children: [
-      { tag: "font-starrail-regular", label: "常规" },
-      { tag: "font-starrail-thin", label: "细体" },
-      { tag: "font-starrail-bold", label: "粗体" },
-    ],
-  },
-  {
-    id: "zzz",
-    label: "绝区零",
-    icon: ImgLogoZzz,
-    displayWidgets: ["icon"],
-    children: [
-      { tag: "font-zzz-a", label: "变体一" },
-      { tag: "font-zzz-b", label: "变体二" },
-    ],
-  },
-];
 
 const { config } = useConfig();
+const { fontOptions } = useFont();
 
 const fontGroupKey = ref("genshin");
 
@@ -65,8 +21,8 @@ const changeFontGroup = (groupKey: string) => {
   fontGroupKey.value = groupKey;
 };
 
-const changeFont = (fontTag: string) => {
-  config.value.font = fontTag;
+const changeFont = (fontOpt: FontNode) => {
+  config.value.font = fontOpt;
 };
 </script>
 
@@ -80,7 +36,7 @@ const changeFont = (fontTag: string) => {
             <el-dropdown-item
               v-for="font in fontGroupList"
               :key="font.tag"
-              :command="font.tag"
+              :command="font"
             >
               {{ font.abbr || font.label }}
             </el-dropdown-item>
