+++ src/views/pageTypewriter/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from "./useConfig";
+export * from "./useFont";
 export * from "./useTextInput";
