+++ src/views/pageTypewriter/hooks/useConfig.ts
@@ -1,9 +1,12 @@
 import type { Config, ConfigInputMethod } from "@/shared";
+import { useFont } from "./useFont";
+
+const { fontDefaultOption } = useFont();
 
 const config: Ref<Config> = ref({
   inputMethod: "keyboard" as ConfigInputMethod,
   keyTransform: true,
-  font: "font-teyvat-regular",
+  font: fontDefaultOption,
 });
 
 export const useConfig = () => {
